Move algorithms to src directory
