--- InsertionSort.java
@@ -1,33 +1,0 @@
-// This program tests the insertion sort algorithm.
-
-// TODO: document running time
-
-import java.util.*; // for Random
-
-/**
- * This program consists of an implementation of the insertion sort algorithm
- * and some examples.
- *
- * TODO: implement algorithm
- *
- * @author Spencer Hubbard
- *
- */
-public class InsertionSort {
-
-  public static void main(String[] args) {
-    // TODO: examples
-  }
-
-  // TODO: comment
-  public static void insertionSort(int[] a) {
-    // TODO: implement
-    throw new RuntimeException("method not implemented");
-  }
-
-  // Helper method.
-  private static void insert(int[] a, int index, int value) {
-    // TODO: implement
-    throw new RuntimeException("method not implemented");
-  }
-}
