+++ src/main/java/com/github/hubbards/algorithms/graph/Color.java
@@ -1,0 +1,10 @@
+package com.github.hubbards.algorithms.graph;
+
+/**
+ * Color used for bookkeeping in graph algorithms.
+ *
+ * @author Spencer Hubbard
+ */
+public enum Color {
+    WHITE, GRAY, BLACK
+}
