+++ src/main/java/com/github/hubbards/algorithms/sort/InsertionSort.java
@@ -1,0 +1,24 @@
+package com.github.hubbards.algorithms.sort;
+
+/**
+ * This class contains an implementation of the insertion sort algorithm.
+ *
+ * @author Spencer Hubbard
+ */
+public final class InsertionSort {
+    private InsertionSort() {
+        throw new AssertionError("static class");
+    }
+
+    // TODO: comment
+    public static void insertionSort(int[] a) {
+        // TODO: implement
+        throw new RuntimeException("method not implemented");
+    }
+
+    // Helper method.
+    private static void insert(int[] a, int index, int value) {
+        // TODO: implement
+        throw new RuntimeException("method not implemented");
+    }
+}
