+++ src/main/java/com/github/hubbards/algorithms/sort/MergeSort.java
@@ -1,26 +1,22 @@
-// This program contains an implementation of the merge sort algorithm.
-
-// NOTE: the worst-case running time of the merge sort algorithm is
-// O(n * log(n)).
+package com.github.hubbards.algorithms.sort;
 
 import java.util.*; // for Arrays
 
-public class MergeSort {
-
-    public static void main(String[] args) {
-        // simple example
-      int a[] = {12, 123, 1, 28, 183, 16};
-      System.out.println("before : " + Arrays.toString(a));
-      mergeSort(a);
-      System.out.println("after  : " + Arrays.toString(a));
-
-        // TODO: test merge sort algorithm
+/**
+ * This class contains an implementation of the merge sort algorithm.
+ *
+ * NOTE: the worst-case running time of the merge sort algorithm is
+ * O(n * log(n)).
+ *
+ * @author Spencer Hubbard
+ */
+public final class MergeSort {
+    private MergeSort() {
+        throw new AssertionError("static class");
     }
 
     // recursive merge sort algorithm
     public static void mergeSort(int[] a) {
-        // print at start of mergeSort method for testing
-        System.out.println("sorting " + Arrays.toString(a));
         if (a.length > 1) {
             // split array into two halves
             int[] left = Arrays.copyOfRange(a, 0, a.length / 2);
@@ -36,8 +32,6 @@ public class MergeSort {
     // pre : result is empty; left and right are sorted
     // post:
     private static void merge(int[] result, int[] left, int[] right) {
-        // print at start of merge method for testing
-        System.out.println("merging " + Arrays.toString(left) + " and " + Arrays.toString(right));
         int i1 = 0; // index into left array
         int i2 = 0; // index into right array
         for (int i = 0; i < result.length; i++) {
