+++ src/test/java/com/github/hubbards/algorithms/graph/SearchGraphTest.java
@@ -1,0 +1,42 @@
+package com.github.hubbards.algorithms.graph;
+
+public class SearchGraphTest {
+    // TODO: convert to unit tests
+    public static void main(String[] args) {
+        SearchGraph g = new SearchGraph();
+        g.addVertex("a");
+        g.addVertex("b");
+        g.addVertex("c");
+        g.addVertex("d");
+        g.addVertex("e");
+        g.addVertex("f");
+        g.addVertex("g");
+        g.addVertex("h");
+        g.addEdge("a", "b");
+        g.addEdge("a", "c");
+        g.addEdge("b", "c");
+        g.addEdge("b", "d");
+        g.addEdge("b", "e");
+        g.addEdge("c", "e");
+        g.addEdge("c", "g");
+        g.addEdge("c", "h");
+        g.addEdge("d", "e");
+        g.addEdge("e", "f");
+        g.addEdge("g", "h");
+
+        System.out.println("Graph:");
+        g.printAdjacencyList();
+        System.out.println();
+
+        System.out.println("BFS from vertex a:");
+        g.breadthFirstSearch("a");
+        System.out.println();
+
+        System.out.println("Iterative DFS from vertex a:");
+        g.depthFirstSearchI("a");
+        System.out.println();
+
+        System.out.println("Recursive DFS from vertex a:");
+        g.depthFirstSearchR("a");
+    }
+}
